Restructure a little bit more
